add test env
