+++ ticketing/auth/Dockerfile
@@ -2,7 +2,7 @@ FROM node:alpine
 
 WORKDIR /app
 COPY package.json .
-RUN npm install
+RUN npm install --omit=dev
 COPY . .
 
 CMD ["npm", "start"]
